move code around a bit; create different service, correct way of creating ingress file
